Benchmarks: Table-fy
Had to un-indent text for the GFM markdown to accept the table. Not sure what to do with rest of the file, so I'll leave it open.

`benchmark/` probably needs some love if we expect it to generate
similar output. Even without it, I still think the most readable for end-users.
is the best possible result.