added set() and get() methods for settings in a koa app
I wanted to be able to do
app.set('port', 3000);
and access it with 
app.get('port')

The key/value pairs are stored in this.settings to avoid naming conflicts. I also modified the original listen to set the port explicitly so that it can be accessed later. If its good, perhaps we can do the same for the other arguments?