force co = 3.0.2
I have co 3.0.2 both as my dependency and as a koa dependency. 
While trying to upgrade my project dependencies (to see if there was any improvement) I upgraded to co-3.0.4 which totally broke anything I then had to go back to 3.0.2 .

I also upgraded koa to 0.5.1, but doing so pulled ```co-0.3.4``` as a koa dependecy breaking again anything.

So it seems to me that there is a “compatibility” issue between co-0.3.2 and c0-0.3.4 that prevents some koa installation (e.g. mine) to correctly work. It would be great if you can accept my pull request so i don't have to do this by hand or i don't have to put this fork as a dependency :)