[CI] Run tests on the stable version of Node
This ensures that Koa is regularly tested against the latest version of Node and surfaces issues early. The main thing is that new version of Node could be released and tests that were previously passing would suddenly start failing (if there were a breaking change).

I also reversed the order of Node versions in the test matrix so that recent = more relevant versions are at the top.