list node versions that Koa recommends
The purpose of this pull is merely a doc fix and matter of morality and conduct.
Would you `require Node v7.6.0 or higher` because it supports async/await but has vulnerabilities prior to the latest and users might end up getting affected in the future, or would you `recommend Node v7.10.1 or v 8.1.4` which also supports async/await and has fixed previous vulnerabilities? 

The doc says it "requires" 7.6.0 or higher, and yet further below `If you're not using node v7.6+, we recommend setting up babel with babel-preset-env`. The 'require' is something that a user 'must' do or have and is not optional. I know many still uses v4 with babel, So i think it is best to list the versions in which Koa 'recommends', with security in mind.