use the ability of `content-type` lib directly
As we can see the doc of `content-type`: [https://github.com/jshttp/content-type#contenttypeparsereq](https://github.com/jshttp/content-type#contenttypeparsereq), it can deal with the original `req` directly and will also throw the exception when `Content-Type` is missing or unvalid. So in my opinion, we don't need `let type = this.get('Content-Type');` to get the type and `if (!type) return ''` to check if the type is unvalid.