Update koa-vs-express.md
As it was, this table made it look like Koa doesn't support routing, templating, sending files, jsonp, which is technically true, but there is Koa middleware built for each of those, so it's helpful to clarify that in the table, rather than leaving those cells empty. It paints Koa in a better light for those who are unfamiliar and are comparing their options, which is the target audience of this page.